Rename to Turnip
CollosalVegetable is a very long name. Turnip is easier to type and
a great reference to the old name.
