--- bin/cli
@@ -1,6 +1,0 @@
-#!/usr/bin/env ruby
-# frozen_string_literal: true
-
-require_relative "../lib/colossal_vegetable"
-
-ColossalVegetable::CLI.new.run
