--- lib/colossal_vegetable.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../lib/colossal_vegetable/cli"
-require_relative "../lib/colossal_vegetable/options"
