+++ lib/turnip.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "../lib/turnip/cli"
+require_relative "../lib/turnip/options"
